requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
fastembed
vanna[snowflake]
openai
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pandas
numpy
azure-core
azure-search
azure-search-documents
